clean file
