+++ public/index.html
@@ -9,7 +9,6 @@
       name="description"
       content="Web site created using create-react-app"
     />
-
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
     <title>Redux Memo</title>
   </head>
