+++ src/containers/AddForm.js
@@ -1,9 +1,9 @@
 import React, { Component } from 'react';
 import styled from 'styled-components';
 import { connect } from 'react-redux';
-import { addMemo } from '../../store/Store';
-import { colors, shadow } from '../../styles/variable';
-import { device } from '../../styles/index';
+import { addMemo } from '../actions/index';
+import { colors, shadow } from '../styles/variable';
+import { device } from '../styles/index';
 
 const Container = styled.div``;
