+++ src/containers/Memo.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import styled from 'styled-components';
 import Item from './Item';
-import { device } from '../../styles/index';
+import { device } from '../styles/index';
 
 const Table = styled.table`
   margin: 0 auto;
