+++ src/index.js
@@ -8,18 +8,18 @@ import { PersistGate } from 'redux-persist/integration/react';
 import { memoReducer } from './store/Store';
 import App from './App';
 
-// Redux Persist Setting
+// Redux-Rersist setting
 
 const persistConfig = {
   key: 'memo',
   storage,
 };
 
-// persist reducer
+// Rersist Reducer
 
 const persistedReducer = persistReducer(persistConfig, memoReducer);
 
-// store persister
+// Store Persister
 
 let store = createStore(persistedReducer);
 let pstore = persistStore(store);
